requirements.txt
docopt==0.6.2
paho-mqtt==1.5.1
passlib==1.7.2
PyYAML==5.3.1
six==1.15.0
transitions==0.8.3
websockets==8.1
